backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
pycryptodome==3.20.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1